+++ backend/src/main/java/br/edu/gamevault/model/Review.java
@@ -2,6 +2,6 @@ package br.edu.gamevault.model;
 
 import java.time.LocalDate;
 
-public record Review(int id, int userId, int gameId, Double rating, String comment, LocalDate reviewDate) {
+public record Review(int id, int userId, int gameId, Integer rating, String comment, LocalDate reviewDate) {
 
 }
